New ebrains-authenticate command
This hopefully resolves the long-standing authentication issue for good.

It is built on the `ebrains-kg-core` library that (among other things)
implements the "device flow" authentication that seems most suitable for
the purpose and target audience of this package.

In the standard case `datalad ebrains-authenticate` is called without
any arguments. It presents instructions to the terminal with a URL to
visit for authentication using a browser. Once successful, it returns
a result record with a token.

A default result renderer is in place to enable shell-based usage ala

```
export KG_TOKEN=`datalad ebrains-authenticate`
```

that puts the token into the `KG_TOKEN` environment variable. This
variable is recognized by the `ebrains-kg-core` itself too, and enabled
token re-use with the following pattern:

```py
>>> from kg_core.kg import kg
>>> k=kg()
>>> k.with_token()
>>> kg_client= k.build()
>>> kg_client.instances.list("https://openminds.ebrains.eu/core/Person")
```

Closes #30