Cloning a `DatasetVersion` UUID now produces this version
Previously all versions known to the corresponding `Dataset` were
discovered and the resulting DataLad dataset had the target version
as one of potentially many version.

This is undesirable (a user likely wants that specific version), and
wasteful (there may be many more versions after the target version).

Now only version *up to* the target version are queried and populated.

Closes #49