Support for shallow clones
This feature is instrumental for export one version, and one version
only, from the knowledge graph. Such an export would be sufficient
for most uses of a reproducible one-time data dependency -- and also
much faster to export for datasets with many versions.

The `--depth` parameter is inspired by that of `git-clone`.

Closes #53