Rough outline of a direct data-proxy query to get bucket content
Previously we went via the KG to get a per-dataset file listing.  This
is extremely slow
(https://github.com/datalad/datalad-ebrains/issues/52).

With this change, we query the data-proxy directly, if a data proxy
bucket is detected to be a datasets's file repository.

This speeds up operations by about 30x.